+++ examples/callback_on_end.rs
@@ -3,11 +3,11 @@ use std::sync::atomic::{AtomicU32, Ordering};
 use std::sync::Arc;
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let stream_handle = rodio::OutputStreamBuilder::open_default_stream()?;
-    let sink = rodio::Sink::connect_new(stream_handle.mixer());
+    let stream_handle = rodio::DeviceSinkBuilder::open_default_sink()?;
+    let player = rodio::Player::connect_new(stream_handle.mixer());
 
     let file = std::fs::File::open("assets/music.wav")?;
-    sink.append(rodio::Decoder::try_from(file)?);
+    player.append(rodio::Decoder::try_from(file)?);
 
     // lets increment a number after `music.wav` has played. We are going to use atomics
     // however you could also use a `Mutex` or send a message through a `std::sync::mpsc`.
@@ -17,7 +17,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     // playlist_pos into the closure. That way we can still access playlist_pos
     // after appending the EmptyCallback.
     let playlist_pos_clone = playlist_pos.clone();
-    sink.append(rodio::source::EmptyCallback::new(Box::new(move || {
+    player.append(rodio::source::EmptyCallback::new(Box::new(move || {
         println!("empty callback is now running");
         playlist_pos_clone.fetch_add(1, Ordering::Relaxed);
     })));
@@ -27,7 +27,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         "playlist position is: {}",
         playlist_pos.load(Ordering::Relaxed)
     );
-    sink.sleep_until_end();
+    player.sleep_until_end();
     assert_eq!(playlist_pos.load(Ordering::Relaxed), 1);
     println!(
         "playlist position is: {}",