+++ examples/music_ogg.rs
@@ -1,14 +1,14 @@
 use std::{error::Error, io::Cursor};
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let stream_handle = rodio::OutputStreamBuilder::open_default_stream()?;
-    let sink = rodio::Sink::connect_new(stream_handle.mixer());
+    let stream_handle = rodio::DeviceSinkBuilder::open_default_sink()?;
+    let player = rodio::Player::connect_new(stream_handle.mixer());
 
     let file = include_bytes!("../assets/music.ogg");
     let cursor = Cursor::new(file);
-    sink.append(rodio::Decoder::try_from(cursor)?);
+    player.append(rodio::Decoder::try_from(cursor)?);
 
-    sink.sleep_until_end();
+    player.sleep_until_end();
 
     Ok(())
 }